decouple path finding
