+++ tests/test_cli/test_lint.py
@@ -1,11 +1,8 @@
 # built-in
 from pathlib import Path
 
-# external
-import pytest
-
 # project
-from deal._cli._lint import get_errors, get_paths, lint_command
+from deal._cli._lint import get_errors, lint_command
 
 
 TEXT = """
@@ -15,23 +12,6 @@ import deal
 def f(x):
     return -1
 """
-
-
-def test_get_paths(tmp_path: Path):
-    (tmp_path / 'subdir').mkdir()
-    (tmp_path / 'subdir' / '__pycache__').mkdir()
-    (tmp_path / '.hidden').mkdir()
-
-    (tmp_path / 'setup.py').touch()
-    (tmp_path / 'subdir' / 'ex.py').touch()
-    (tmp_path / '.hidden' / 'ex.py').touch()
-    (tmp_path / 'subdir' / '__pycache__' / 'ex.py').touch()
-    (tmp_path / 'setup.pl').touch()
-    actual = {p.relative_to(tmp_path) for p in get_paths(tmp_path)}
-    assert actual == {Path('setup.py'), Path('subdir/ex.py')}
-
-    with pytest.raises(FileNotFoundError):
-        list(get_paths(tmp_path / 'not_exists'))
 
 
 def test_get_errors(tmp_path: Path):
